+++ prisma.config.ts
@@ -1,9 +1,10 @@
+import 'dotenv/config';
 import path from 'node:path';
-import { defineConfig } from 'prisma/config';
+import { defineConfig, env } from 'prisma/config';
 
 export default defineConfig({
   schema: path.join('prisma', 'schema.prisma'),
   datasource: {
-    url: 'file:./dev.db',
+    url: env('DATABASE_URL'),
   },
 });